% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CanekDebug-package.R
\docType{package}
\name{CanekDebug-package}
\alias{CanekDebug}
\alias{CanekDebug-package}
\title{CanekDebug: Functions to Evaluate Canek Integration}
\description{
A set of functions to evaluate different aspects of Canek 
    integration.
}
\author{
\strong{Maintainer}: Diego Diez \email{diego10ruiz@gmail.com} (\href{https://orcid.org/0000-0002-2325-4893}{ORCID})

}
\keyword{internal}
