% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Membership.R
\name{CheckUncorrectedMemberships}
\alias{CheckUncorrectedMemberships}
\title{CheckUncorrectedMemberships}
\usage{
CheckUncorrectedMemberships(x)
}
\arguments{
\item{x}{CanekDebug object.}
}
\description{
CheckUncorrectedMemberships
}
