% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debug.R
\name{GetBatchCellNames}
\alias{GetBatchCellNames}
\title{GetBatchCellNames}
\usage{
GetBatchCellNames(x, batch = 1)
}
\arguments{
\item{x}{CanekDebug object.}

\item{batch}{batch correction number/name.}
}
\description{
GetBatchCellNames
}
