% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correction.R
\name{GetCorrectionMatrix}
\alias{GetCorrectionMatrix}
\title{GetCorrectionMatrix}
\usage{
GetCorrectionMatrix(x, batch = 1)
}
\arguments{
\item{x}{CanekDebug object.}

\item{batch}{batch correction number/name.}
}
\description{
GetCorrectionMatrix
}
