% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Membership.R
\name{GetMembershipIndex}
\alias{GetMembershipIndex}
\title{GetMembershipIndex}
\usage{
GetMembershipIndex(x, batch = 1)
}
\arguments{
\item{x}{CanekDebug object.}

\item{batch}{batch correction number/name.}
}
\description{
GetMembershipIndex
}
