% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debug.R
\name{IntersectFeatures}
\alias{IntersectFeatures}
\title{IntersectFeatures}
\usage{
IntersectFeatures(x, assay = "RNA", slot = "data", use.hvf = TRUE, cutoff = 0)
}
\arguments{
\item{x}{Seurat object.}

\item{assay}{assay name.}

\item{slot}{slot name.}

\item{use.hvf}{whether to use HVF.}

\item{cutoff}{cutoff for gene expression.}
}
\description{
IntersectFeatures
}
