% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotBatchMembership}
\alias{plotBatchMembership}
\title{plotBatchMembership}
\usage{
plotBatchMembership(
  x,
  batch = 1,
  reduction = "umapraw",
  add.ref = TRUE,
  plot.membership = TRUE,
  plot.mnn = FALSE,
  ...
)
}
\arguments{
\item{x}{CanekDebug object.}

\item{batch}{batch correction number/name.}

\item{reduction}{reduction to plot}

\item{add.ref}{whether to plot reference cells.}

\item{plot.membership}{whether to color by memberships.}

\item{plot.mnn}{whether to add MNN pairs.}

\item{...}{arguments passed down to methods.}
}
\description{
plotBatchMembership
}
