% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debug.R
\name{plotCorrectionHist}
\alias{plotCorrectionHist}
\title{plotCorrectionHist}
\usage{
plotCorrectionHist(x, batch = 1, by.membership = TRUE)
}
\arguments{
\item{x}{CanekDebug object.}

\item{batch}{batch correction number/name.}

\item{by.membership}{whether facet plot by membership.}
}
\description{
plotCorrectionHist
}
