% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotNumberOfCorrectedGenes}
\alias{plotNumberOfCorrectedGenes}
\title{plotNumberOfCorrectedGenes}
\usage{
plotNumberOfCorrectedGenes(x)
}
\arguments{
\item{x}{CanekDebug object.}
}
\description{
plotNumberOfCorrectedGenes
}
