% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotTopGenes}
\alias{plotTopGenes}
\title{plotTopGenes}
\usage{
plotTopGenes(x, batch = NULL)
}
\arguments{
\item{x}{CanekDebug object.}

\item{batch}{batch correction number/name.}
}
\description{
plotTopGenes
}
