% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debug.R
\name{SummarizeTopGenes}
\alias{SummarizeTopGenes}
\title{SummarizeTopGenes}
\usage{
SummarizeTopGenes(x, batch = NULL)
}
\arguments{
\item{x}{CanekDebug object.}

\item{batch}{batch correction number/name.}
}
\description{
SummarizeTopGenes
}
